requirements.txt
fastapi>=0.100.0,<0.110.0
# uvicorn is installed separately in the Dockerfile
# pydantic is installed separately in the Dockerfile
sqlalchemy>=2.0.0,<2.1.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.1.0
aiosqlite>=0.17.0,<0.20.0
alembic>=1.10.0,<1.13.0
python-dotenv>=0.21.0,<1.1.0